requirements.txt
langchain
ipykernel
python-dotenv
langchain_community
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
langchain-openai
chromadb
langchain-chroma
langchain-milvus
pgvector
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
langchain-perplexity
beautifulsoup4
langchain_google_genai
langchain_perplexity
streamlit
langchain-ollama
langchain-groq
fastapi 
uvicorn
langserve
sse_starlette